+++ src/main/java/svenhjol/charmony/core/base/Config.java
@@ -125,11 +125,12 @@ public final class Config {
             for (var sided : feature.sides()) {
                 var config = configs.get(sided.side());
 
+                //noinspection deprecation
                 if (!sided.showInConfig()) {
-                    continue;
+                    continue; // TODO: This is deprecated behavior and mods need to move away from using it.
                 }
 
-                if (sided.canBeDisabled()) {
+                if (sided.canBeDisabled() && sided.canBeDisabledInConfig()) {
                     var field = "Enabled";
                     var description = sided.description();
                     var configName = sided.className() + "." + field;